pyproject.toml
[project]
name = "taxonomy2"
version = "0.1.0"
description = "Taxonomy Discovery & Entity Resolution Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "litellm",
    "pymupdf",
    "python-docx",
    "openpyxl",
    "numpy",
    "python-multipart",
    "pydantic",
    "pydantic-settings",
    "pillow",
    "xlrd>=2.0.2",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio>=1.3.0",
    "xlwt>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
